Extract Fleet setup
