+++ x-pack/test_serverless/api_integration/test_suites/observability/fleet/fleet.ts
@@ -7,48 +7,18 @@
 
 import expect from 'expect';
 import { FtrProviderContext } from '../../../ftr_provider_context';
+import {
+  expectDefaultElasticsearchOutput,
+  expectDefaultFleetServer,
+} from '../../common/fleet/default_setup';
 
-export default function ({ getService }: FtrProviderContext) {
-  const svlCommonApi = getService('svlCommonApi');
-  const supertest = getService('supertest');
-  const retry = getService('retry');
-
-  const defaultFleetServerHostId = 'default-fleet-server';
-  const defaultFleetServerHostUrl = 'https://localhost:8220';
-  const defaultElasticsearchOutputId = 'es-default-output';
-  const defaultElasticsearchOutputHostUrl = 'https://localhost:9200';
-
-  async function expectDefaultFleetServer() {
-    await retry.waitForWithTimeout('get default fleet server', 30_000, async () => {
-      const { body, status } = await supertest.get(
-        `/api/fleet/fleet_server_hosts/${defaultFleetServerHostId}`
-      );
-      if (status === 200 && body.item.host_urls.includes(defaultFleetServerHostUrl)) {
-        return true;
-      } else {
-        throw new Error(`Expected default Fleet Server id ${defaultFleetServerHostId} to exist`);
-      }
-    });
-  }
-
-  async function expectDefaultElasticsearchOutput() {
-    await retry.waitForWithTimeout('get default Elasticsearch output', 30_000, async () => {
-      const { body, status } = await supertest.get(
-        `/api/fleet/outputs/${defaultElasticsearchOutputId}`
-      );
-      if (status === 200 && body.item.hosts.includes(defaultElasticsearchOutputHostUrl)) {
-        return true;
-      } else {
-        throw new Error(
-          `Expected default Elasticsearch output id ${defaultElasticsearchOutputId} to exist`
-        );
-      }
-    });
-  }
+export default function (ctx: FtrProviderContext) {
+  const svlCommonApi = ctx.getService('svlCommonApi');
+  const supertest = ctx.getService('supertest');
 
   describe('fleet', function () {
     it('rejects request to create a new fleet server hosts if host url is different from default', async () => {
-      await expectDefaultFleetServer();
+      await expectDefaultFleetServer(ctx);
 
       const { body, status } = await supertest
         .post('/api/fleet/fleet_server_hosts')
@@ -62,13 +32,13 @@ export default function ({ getService }: FtrProviderContext) {
       expect(body).toEqual({
         statusCode: 403,
         error: 'Forbidden',
-        message: `Fleet server host must have default URL in serverless: ${defaultFleetServerHostUrl}`,
+        message: 'Fleet server host must have default URL in serverless: https://localhost:8220',
       });
       expect(status).toBe(403);
     });
 
     it('accepts request to create a new fleet server hosts if host url is same as default', async () => {
-      await expectDefaultFleetServer();
+      await expectDefaultFleetServer(ctx);
 
       const { body, status } = await supertest
         .post('/api/fleet/fleet_server_hosts')
@@ -81,14 +51,14 @@ export default function ({ getService }: FtrProviderContext) {
       expect(body).toEqual({
         item: expect.objectContaining({
           name: 'Test Fleet server host',
-          host_urls: [defaultFleetServerHostUrl],
+          host_urls: ['https://localhost:8220'],
         }),
       });
       expect(status).toBe(200);
     });
 
     it('rejects request to create a new elasticsearch output if host is different from default', async () => {
-      await expectDefaultElasticsearchOutput();
+      await expectDefaultElasticsearchOutput(ctx);
 
       const { body, status } = await supertest
         .post('/api/fleet/outputs')
@@ -103,13 +73,14 @@ export default function ({ getService }: FtrProviderContext) {
       expect(body).toEqual({
         statusCode: 400,
         error: 'Bad Request',
-        message: `Elasticsearch output host must have default URL in serverless: ${defaultElasticsearchOutputHostUrl}`,
+        message:
+          'Elasticsearch output host must have default URL in serverless: https://localhost:9200',
       });
       expect(status).toBe(400);
     });
 
     it('accepts request to create a new elasticsearch output if host url is same as default', async () => {
-      await expectDefaultElasticsearchOutput();
+      await expectDefaultElasticsearchOutput(ctx);
 
       const { body, status } = await supertest
         .post('/api/fleet/outputs')
@@ -123,7 +94,7 @@ export default function ({ getService }: FtrProviderContext) {
       expect(body).toEqual({
         item: expect.objectContaining({
           name: 'Test output',
-          hosts: [defaultElasticsearchOutputHostUrl],
+          hosts: ['https://localhost:9200'],
         }),
       });
       expect(status).toBe(200);
