+++ x-pack/test_serverless/api_integration/test_suites/security/fleet/config.ts
@@ -6,6 +6,7 @@
  */
 
 import { createTestConfig } from '../../../config.base';
+import { kbnServerArgs } from '../../common/fleet/default_setup';
 
 export default createTestConfig({
   serverlessProject: 'security',
@@ -19,21 +20,5 @@ export default createTestConfig({
   // https://github.com/elastic/project-controller/blob/main/internal/project/security/config/elasticsearch.yml
   esServerArgs: ['xpack.ml.nlp.enabled=false'],
 
-  kbnServerArgs: [
-    '--xpack.cloud.serverless.project_id=ftr_fake_project_id',
-    `--xpack.fleet.fleetServerHosts=[${JSON.stringify({
-      id: 'default-fleet-server',
-      name: 'Default Fleet Server',
-      is_default: true,
-      host_urls: ['https://localhost:8220'],
-    })}]`,
-    `--xpack.fleet.outputs=[${JSON.stringify({
-      id: 'es-default-output',
-      name: 'Default Output',
-      type: 'elasticsearch',
-      is_default: true,
-      is_default_monitoring: true,
-      hosts: ['https://localhost:9200'],
-    })}]`,
-  ],
+  kbnServerArgs,
 });
